+++ services/supervisor/src/transactions.rs
@@ -22,7 +22,7 @@ use exonum_merkledb::ObjectHash;
 use std::collections::HashSet;
 
 use super::{
-    mode, ConfigChange, ConfigProposalWithHash, ConfigPropose, ConfigVote, ConfigureCall,
+    ConfigChange, ConfigProposalWithHash, ConfigPropose, ConfigVote, ConfigureCall,
     DeployConfirmation, DeployRequest, Error, Schema, StartService, Supervisor,
 };
 
@@ -118,10 +118,7 @@ impl StartService {
     }
 }
 
-impl<Mode> SupervisorInterface for Supervisor<Mode>
-where
-    Mode: mode::SupervisorMode,
-{
+impl SupervisorInterface for Supervisor {
     fn propose_config_change(
         &self,
         mut context: CallContext<'_>,
@@ -285,7 +282,8 @@ where
         }
 
         schema.deploy_requests.confirm(&deploy, author);
-        if Mode::deploy_approved(&deploy, &schema.deploy_requests, validator_count) {
+        let supervisor_mode = schema.supervisor_config().mode;
+        if supervisor_mode.deploy_approved(&deploy, &schema.deploy_requests, validator_count) {
             log::trace!("Deploy artifact request accepted {:?}", deploy.artifact);
             let artifact = deploy.artifact.clone();
             schema.pending_deployments.put(&artifact, deploy);
@@ -345,10 +343,7 @@ where
     }
 }
 
-impl<Mode> Supervisor<Mode>
-where
-    Mode: mode::SupervisorMode,
-{
+impl Supervisor {
     /// Verifies that each change introduced within config proposal is valid.
     fn verify_config_changeset(
         &self,